+++ src/client/components/PageChantier/ComparaisonTerritoires/ComparaisonTerritoires.tsx
@@ -1,22 +1,9 @@
-import { useCallback } from "react";
 import { MailleInterne } from "@/server/domain/maille/Maille.interface";
 import { WidgetCartographieTA } from "@/components/_commons/Widget/WidgetCartographieTA/WidgetCartographieTA";
 import { WidgetCartographieMeteo } from "@/components/_commons/Widget/WidgetCartographieMeteo/WidgetCartographieMeteo";
 import { WidgetCartographiePVA } from "@/components/_commons/Widget/WidgetCartographiePVA/WidgetCartographiePVA";
 import { ComparaisonTerritoires as ComparaisonTerritoiresBase } from "@/components/_commons/ComparaisonTerritoires/ComparaisonTerritoires";
-import { getLabelTerritoire } from "@/client/constants/territoires";
-import { useTerritoiresCompares } from "@/client/hooks/useTerritoiresCompares";
-import {
-  ContenuCsv,
-  filtrerTerritoires,
-  formaterDateCsv,
-  genererCsv,
-  telechargerCsv,
-} from "@/client/utils/csv/genererCsv";
-import api from "@/server/infrastructure/api/trpc/api";
-import { TauxAvancementComparaisonTerritoireViewModel } from "@/server/chantiers/app/contrats/TauxAvancementComparaisonTerritoireViewModel";
-import { MeteoTerritoireViewModel } from "@/server/chantiers/infrastructure/queries/GetChantierMeteosTerritoiresQuery";
-import { PVATerritoireViewModel } from "@/server/chantiers/infrastructure/queries/GetChantierPVACountTerritoiresQuery";
+import { useExporterComparaisonEnCsv } from "./useExporterComparaisonEnCsv";
 
 type ComparaisonTerritoiresProps = {
   chantierId: string;
@@ -25,7 +12,7 @@ type ComparaisonTerritoiresProps = {
   territoireCode: string;
 };
 
-type TypeCarteChantier = "ta" | "meteo" | "pva";
+export type TypeCarteChantier = "ta" | "meteo" | "pva";
 
 const options: (
   jalon: number,
@@ -38,129 +25,17 @@ const options: (
   },
 ];
 
-export const construireContenuCsv = (
-  params:
-    | {
-        type: "ta";
-        donneesTA: TauxAvancementComparaisonTerritoireViewModel[];
-        codesTerritoiresSelectionnes: string[];
-      }
-    | {
-        type: "meteo";
-        donneesMeteo: MeteoTerritoireViewModel[];
-        codesTerritoiresSelectionnes: string[];
-      }
-    | {
-        type: "pva";
-        donneesPVA: PVATerritoireViewModel[];
-        codesTerritoiresSelectionnes: string[];
-      },
-): ContenuCsv => {
-  if (params.type === "ta") {
-    const { donneesTA, codesTerritoiresSelectionnes } = params;
-    return {
-      colonnes: [
-        "Territoire",
-        "Taux d'avancement",
-        "Date de dernière mise à jour",
-      ],
-      lignes: filtrerTerritoires(donneesTA, codesTerritoiresSelectionnes).map(
-        (territoire) => [
-          getLabelTerritoire(territoire.territoireCode),
-          territoire.tauxAvancementJalon !== null
-            ? String(territoire.tauxAvancementJalon)
-            : "Non renseigné",
-          formaterDateCsv(territoire.dateTauxAvancementAnnuel),
-        ],
-      ),
-    };
-  }
-
-  if (params.type === "meteo") {
-    const { donneesMeteo, codesTerritoiresSelectionnes } = params;
-    return {
-      colonnes: ["Territoire", "Niveau de confiance", "Date de publication"],
-      lignes: filtrerTerritoires(
-        donneesMeteo,
-        codesTerritoiresSelectionnes,
-      ).map((territoire) => [
-        getLabelTerritoire(territoire.territoireCode),
-        territoire.meteo,
-        formaterDateCsv(territoire.dateDeMajQualitative),
-      ]),
-    };
-  }
-
-  const { donneesPVA, codesTerritoiresSelectionnes } = params;
-  return {
-    colonnes: ["Territoire", "Nombre de propositions"],
-    lignes: filtrerTerritoires(donneesPVA, codesTerritoiresSelectionnes).map(
-      (territoire) => [
-        getLabelTerritoire(territoire.territoireCode),
-        String(territoire.nombrePropositionsValeur),
-      ],
-    ),
-  };
-};
-
 export const ComparaisonTerritoires = ({
   chantierId,
   jalon,
   maille,
   territoireCode,
 }: ComparaisonTerritoiresProps) => {
-  const utils = api.useUtils();
-  const [territoiresCompares] = useTerritoiresCompares();
-
-  const exporterEnCsv = useCallback(
-    (type: TypeCarteChantier) => {
-      const codesTerritoiresSelectionnes = [
-        territoireCode,
-        ...territoiresCompares.split(",").filter(Boolean),
-      ];
-      const nomFichier = `comparaison-territoriale-${chantierId}-${type}.csv`;
-
-      if (type === "ta") {
-        const { colonnes, lignes } = construireContenuCsv({
-          type,
-          donneesTA:
-            utils.chantier.recupererTauxAvancementTerritoires.getData({
-              chantierIds: [chantierId],
-              jalon,
-            }) ?? [],
-          codesTerritoiresSelectionnes,
-        });
-        telechargerCsv(genererCsv(colonnes, lignes), nomFichier);
-        return;
-      }
-
-      if (type === "meteo") {
-        const { colonnes, lignes } = construireContenuCsv({
-          type,
-          donneesMeteo:
-            utils.chantier.recupererMeteosTerritoires.getData({
-              chantierId,
-              jalon,
-            }) ?? [],
-          codesTerritoiresSelectionnes,
-        });
-        telechargerCsv(genererCsv(colonnes, lignes), nomFichier);
-        return;
-      }
-
-      const { colonnes, lignes } = construireContenuCsv({
-        type,
-        donneesPVA:
-          utils.chantier.recupererPVAChantierTerritoires.getData({
-            chantierId,
-            jalon,
-          }) ?? [],
-        codesTerritoiresSelectionnes,
-      });
-      telechargerCsv(genererCsv(colonnes, lignes), nomFichier);
-    },
-    [utils, chantierId, jalon, territoireCode, territoiresCompares],
-  );
+  const exporterEnCsv = useExporterComparaisonEnCsv({
+    chantierId,
+    jalon,
+    territoireCode,
+  });
 
   return (
     <ComparaisonTerritoiresBase<TypeCarteChantier>
